chore: migrate network and tenant state to context
